requirements.txt
numpy==1.18.1
torchvision==0.6.0
pytorch==1.5.1
pyyaml==3.13
tqdm==4.26.0
Pillow==5.3.0
opencv-python